backend/pyproject.toml
[tool.poetry]
name = "waterlooworks-adapter"
version = "0.1.0"
description = ""
authors = ["Krish Shah <shahkrish2016@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
lxml = "^5.2.2"
google-api-python-client = "^2.134.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
toml = "^0.10.2"
google-auth = "^2.30.0"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
